+++ example/historical/metadata.cpp
@@ -13,7 +13,7 @@ int main() {
   const auto publishers = client.MetadataListPublishers();
   std::cout << "Publishers:\n";
   for (const auto& publisher : publishers) {
-    std::cout << "- " << publisher.first << ": " << publisher.second << '\n';
+    std::cout << "- " << publisher << '\n';
   }
   std::cout << '\n';
 
@@ -31,15 +31,11 @@ int main() {
   }
   std::cout << '\n';
 
-  const auto fields = client.MetadataListFields(
-      kGlbxMdp3, databento::Encoding::Dbn, databento::Schema::Trades);
+  const auto fields = client.MetadataListFields(databento::Encoding::Dbn,
+                                                databento::Schema::Trades);
   std::cout << "Fields:\n";
-  const auto& dbn_trades_fields = fields.at("GLBX.MDP3")
-                                      .at(databento::Encoding::Dbn)
-                                      .at(databento::Schema::Trades);
-  for (const auto& field_and_type : dbn_trades_fields) {
-    std::cout << "- " << field_and_type.first << ": " << field_and_type.second
-              << '\n';
+  for (const auto& field_detail : fields) {
+    std::cout << "- " << field_detail << '\n';
   }
   std::cout << '\n';
 
@@ -54,8 +50,8 @@ int main() {
   const auto all_unit_prices = client.MetadataListUnitPrices(kGlbxMdp3);
   std::cout << "Unit prices:\n";
   for (const auto& mode_and_prices : all_unit_prices) {
-    const auto* mode_str = ToString(mode_and_prices.first);
-    for (const auto& schema_and_price : mode_and_prices.second) {
+    const auto* mode_str = ToString(mode_and_prices.mode);
+    for (const auto& schema_and_price : mode_and_prices.unit_prices) {
       std::cout << "- (" << mode_str << ", " << schema_and_price.first
                 << "): " << schema_and_price.second << '\n';
     }
@@ -66,29 +62,6 @@ int main() {
       client.MetadataGetRecordCount(kGlbxMdp3, {"2020-12-28", "2020-12-29"},
                                     {"ESH1"}, databento::Schema::Mbo);
   std::cout << "Record count: " << record_count << "\n\n";
-
-  const auto live_unit_prices =
-      client.MetadataListUnitPrices(kGlbxMdp3, databento::FeedMode::Live);
-  std::cout << "Unit prices (live):\n";
-  for (const auto& schema_and_price : live_unit_prices) {
-    std::cout << "- (" << schema_and_price.first
-              << "): " << schema_and_price.second << '\n';
-  }
-  std::cout << '\n';
-
-  const auto trades_unit_prices =
-      client.MetadataListUnitPrices(kGlbxMdp3, databento::Schema::Trades);
-  std::cout << "Unit prices (trades):\n";
-  for (const auto& mode_and_price : trades_unit_prices) {
-    std::cout << "- (" << mode_and_price.first << "): " << mode_and_price.second
-              << '\n';
-  }
-  std::cout << '\n';
-
-  const auto unit_price = client.MetadataListUnitPrices(
-      kGlbxMdp3, databento::FeedMode::Historical, databento::Schema::Trades);
-  std::cout << "Unit price (GLBX.MDP3, Historical, Trades): " << unit_price
-            << "\n\n";
 
   const std::size_t billable_size = client.MetadataGetBillableSize(
       kGlbxMdp3, {"2020-12-28", "2020-12-29"}, {"ESH1"}, databento::Schema::Mbo,