+++ src/detail/json_helpers.cpp
@@ -72,6 +72,20 @@ std::uint64_t ParseAt(const std::string& endpoint, const nlohmann::json& json,
 }
 
 template <>
+std::uint16_t ParseAt(const std::string& endpoint, const nlohmann::json& json,
+                      const std::string& key) {
+  const auto& val_json = CheckedAt(endpoint, json, key);
+  if (val_json.is_null()) {
+    return 0;
+  }
+  if (!val_json.is_number_unsigned()) {
+    throw JsonResponseError::TypeMismatch(endpoint, key + " unsigned number",
+                                          val_json);
+  }
+  return val_json;
+}
+
+template <>
 double ParseAt(const std::string& endpoint, const nlohmann::json& json,
                const std::string& key) {
   const auto& val_json = CheckedAt(endpoint, json, key);